requirements.txt
google-genai
amadeus

